Update Toast code
Move showToast function to utils.js
